runtime: fix new scheduler on freebsd, windows
R=devon.odell
CC=golang-dev
https://golang.org/cl/7443046

+++ src/pkg/runtime/thread_plan9.c
@@ -228,7 +228,7 @@ runtime·newosproc(M *mp, G *gp, void *stk, void (*fn)(void))
 	mp->tls[0] = mp->id;	// so 386 asm can find it
 	if(0){
 		runtime·printf("newosproc stk=%p m=%p g=%p fn=%p rfork=%p id=%d/%d ostk=%p\n",
-			stk, mp, gp, fn, runtime·rfork, mp->id, mp->tls[0], &mp);
+			stk, mp, gp, fn, runtime·rfork, mp->id, (int32)mp->tls[0], &mp);
 	}
 
 	if(runtime·rfork(RFPROC|RFMEM|RFNOWAIT, stk, mp, gp, fn) < 0)
